LPK-7125 Add BIM model and attachment types
Add BIM model and model attachment types and related translations to
the Rakennuslupa_v2 structure. New BIM models are added under the key
:tietomallit. This includes distinct types for the architectural,
structure, building services and construction site models. The BIM
models support the same archival metadata as the documents.

The model attachment types are added under the :tietomallien_liitteet
key. If multiple BIM models are added under the same parent in the
metadata configuration, the attachments need only be added under one
of the models. The same attachment metadata is applied to the BIM
attachments uploaded to all of the models.

Bumps version number to 5.3.18.